+++ src/app/[locale]/layout.jsx
@@ -1,3 +1,5 @@
+import "@/src/styles/global.scss";
+
 import CookiesModal from "@/src/components/modals/CookiesModal/CookiesModal";
 import PaymentModal from "@/src/components/modals/PaymentModal/PaymentModal";
 import Footer from "@/src/components/shared/Footer/Footer";
@@ -5,8 +7,6 @@ import Header from "@/src/components/shared/Header/Header";
 import { NextIntlClientProvider } from "next-intl";
 import { getMessages } from "next-intl/server";
 import NextTopLoader from "nextjs-toploader";
-import "@/src/styles/global.scss";
-import ErrorAlert from "@/src/components/shared/ErrorAlert/ErrorAlert";
 import ScrollToTopBtn from "@/src/components/shared/ScrollToTopBtn/ScrollToTopBtn";
 
 export const metadata = {
