+++ src/components/ContactFormSection/ContactForm/ContactForm.jsx
@@ -10,14 +10,14 @@ import { FeedbackSchema, feedbackDefaultValues } from "./FeedbackSchema";
 import { Icon } from "../../shared/Icon/Icon";
 import MainButton from "../../shared/MainButton/MainButton";
 import InputField from "../../shared/InputField/InputField";
-import stateErrorAlert from "@/src/state/stateErrorAlert";
-import ErrorAlert from "../../shared/ErrorAlert/ErrorAlert";
+import stateUseAlert from "@/src/state/stateUseAlert";
+import UseAlert from "../../shared/UseAlert/UseAlert";
 import Loader from "../../shared/loader/Loader";
 
 export default function ContactForm() {
   const t = useTranslations("Main.feedback_form");
-  const isOpen = stateErrorAlert(state => state.isOpen);
-  const open = stateErrorAlert(state => state.open);
+  const isOpen = stateUseAlert(state => state.isOpen);
+  const open = stateUseAlert(state => state.open);
 
   const {
     register,
@@ -106,7 +106,7 @@ export default function ContactForm() {
         <Icon name='mail'/>
       </div>}
       
-      {isOpen && <ErrorAlert/>}
+      {isOpen && <UseAlert/>}
 
       {loader && <Loader/>}
     </form>
