+++ src/components/modals/CookiesModal/CookiesModal.module.scss
@@ -5,7 +5,6 @@
   position: fixed;
   left: 50%;
   transform: translateX(-50%);
-  //right: auto;
   bottom: 14%;
   top: auto;
   z-index: 1000;
