+++ src/components/shared/UseAlert/UseAlert.module.scss
@@ -2,11 +2,20 @@
 @use "../../../styles/mixins.scss" as mx;
 
 .wrapper{
+  min-width: 98%;
   padding: 10px;
   position: fixed;
-  left: 0;
-  bottom: 6%;
+  left: 50%;
+  transform: translateX(-50%);
+  bottom: 13%;
   z-index: 20;
+
+  @include mx.media-q("md") {
+    min-width:auto;
+    transform: none;
+    left: 1%;
+    bottom: 8%;
+  }
 
   .alert{
     display: flex;
@@ -15,11 +24,17 @@
     justify-content: center;
     position: relative;
     padding: 24px 16px;
-    color: var.$secondary-error;
+    //color: var.$secondary-error;
     border: 2px solid var.$primary-border;
     border-radius: 12px;
     background: var.$primary-fon-card;
-  
+    &.error{
+      color: var.$secondary-error;
+    }
+    &.success{
+      color: var.$secondary-positive;
+    }
+
     & > h2{
       @include mx.apply-font-styles(var.$h1-font);
       font-size: 20px;
